requirements.txt
pytorch-lightning>=2.4.0
path>=17.0.0
fvcore>=0.1.5.post20221221
timm>=1.0.8
wandb>=0.19.7
torchvision>=0.17.0
pillow>=10.3.0
numpy>=1.26.4